Check these invariants after reconfigurations as well
