+++ tests/partitions_test.py
@@ -16,7 +16,7 @@ import time
 import http
 import contextlib
 import ccf.ledger
-from governance_history import check_signatures
+from reconfiguration import test_ledger_invariants
 
 from loguru import logger as LOG
 
@@ -791,18 +791,6 @@ def test_session_consistency(network, args):
     return network
 
 
-@reqs.description("Confirm ledger contains expected entries")
-def test_ledger_invariants(network, args):
-    # Force ledger flush of all transactions so far
-    network.get_latest_ledger_public_state()
-
-    for node in network.nodes:
-        LOG.info(f"Examining ledger on node {node.local_node_id}")
-        ledger_directories = node.remote.ledger_paths()
-        ledger = ccf.ledger.Ledger(ledger_directories)
-        check_signatures(ledger)
-
-
 def run_2tx_reconfig_tests(args):
     if not args.include_2tx_reconfig:
         return
@@ -849,6 +837,7 @@ def run(args):
         test_election_reconfiguration(network, args)
         test_forwarding_timeout(network, args)
         test_session_consistency(network, args)
+
         test_ledger_invariants(network, args)
